split gui and control so AS state can continue publishing - successfully replaces vehicle supervisor for manual overriding states
